Move configs for status to internal.config.
